Move the unit tests to the correct place
