+++ sunpy/map/tests/test_mapbase.py
@@ -37,90 +37,7 @@ from sunpy.time import parse_time
 from sunpy.util import SunpyUserWarning
 from sunpy.util.exceptions import SunpyDeprecationWarning, SunpyMetadataWarning
 from sunpy.util.metadata import ModifiedItem
-from sunpy.util.util import fix_duplicate_notes
 from .strategies import matrix_meta
-
-
-def test_notes_combined():
-    map_documentation = """
-Class Info.
-
-Notes
------
-This is a note.
-
-References
-----------
-This is reference."""
-    extra_note_section = """\nNotes\n-----\nThis should be combined."""
-    updated_documentation = fix_duplicate_notes(extra_note_section, map_documentation)
-    expected_result = """
-Class Info.
-
-Notes
------
-This is a note.
-
-This should be combined.
-
-References
-----------
-This is reference."""
-    assert updated_documentation.strip() == expected_result.strip()
-
-def test_notes_combined_no_references():
-    map_documentation = """
-Class Info.
-
-Notes
------
-This is a note."""
-    extra_note_section = """\nNotes\n-----\nThis should be combined."""
-    updated_documentation = fix_duplicate_notes(extra_note_section, map_documentation)
-    expected_result = """
-Class Info.
-
-Notes
------
-This is a note.
-
-This should be combined."""
-    assert updated_documentation.strip() == expected_result.strip()
-
-def test_notes_combined_no_existing_notes():
-    map_documentation = """
-Class Info.
-
-References
-----------
-This is reference."""
-    extra_note_section= """\nNotes\n-----\nThis should be combined."""
-    updated_documentation = fix_duplicate_notes(extra_note_section, map_documentation)
-    expected_result = """
-Class Info.
-
-Notes
------
-This should be combined.
-
-References
-----------
-This is reference."""
-    assert updated_documentation.strip() == expected_result.strip()
-
-def test_notes_combined_no_notes_no_references():
-    map_documentation = """
-Class Info.
-"""
-    extra_note_section= """\nNotes\n-----\nThis should be combined."""
-    updated_documentation= fix_duplicate_notes(extra_note_section, map_documentation)
-    expected_result = """
-Class Info.
-
-Notes
------
-This should be combined."""
-    assert updated_documentation.strip() == expected_result.strip()
 
 
 def test_fits_data_comparison(aia171_test_map):
